+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 ### Added
+- Transformations for LiveData:
+  - `mapNotNull`
+  - `filter`
+  - `filterNotNull`
+  - `filterIsInstance`
+  - `reduce`
+  - `reduceNotNull`
+  - `throttleWithTimeout`
 - JUnit5 extensions for testing LiveData:
   - `InstantTaskExecutorExtension`
   - `TestCoroutineDispatcherExtension`